+++ CHANGELOG.md
@@ -7,6 +7,7 @@ This project adheres to [Semantic Versioning](http://semver.org/).
 
 ## Version [2.2.0][unreleased] - (in development)
 - Add: [Locales module](https://www.contentful.com/developers/docs/references/content-management-api/#/reference/locales).
+- Fix: ImageMeta contains data now.
 
 ## Version [2.1.0]- 2017-07-19
 - Add: Custom user agent header.